requirements.txt
sstreamlit
pandas
numpy
yfinance
scikit-learn
tensorflow
matplotlib
